requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# RabbitMQ
pika==1.3.2

# API Documentation
drf-yasg==1.21.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development/Testing
pytest==7.4.3
pytest-django==4.7.0
faker==20.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring (optional)
django-prometheus==2.3.1
